Added more tests and updated some classes to fit the front-end.

+++ ecommerce/src/main/java/com/ecommerce/repositories/implementations/ReviewRepositoryImplementation.java
@@ -45,16 +45,26 @@ public class ReviewRepositoryImplementation implements ReviewRepository {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long userId, Long productId) {
         Session session = sessionFactory.getCurrentSession();
-        Review review = session.get(Review.class, id);
+        User user = session.get(User.class, userId);
+        Product product = session.get(Product.class, productId);
+        Review review = session.get(Review.class,id);
+        if(user==null || product == null || review == null){
+            throw new NotFoundException();
+        }
+
+        Hibernate.initialize(user.getReviews());
+        Hibernate.initialize(product.getReviewList());
+        if(!user.getReviews().contains(review) || !product.getReviewList().contains(review)){
+            throw new NotFoundException();
+        }
+        user.removeReview(review);
+        product.removeReview(review);
+
         session.delete(review);
     }
 
-    @Override
-    public void delete(Review review) {
-        deleteById(review.getId());
-    }
 
     @Override
     public void save(Review review, Long idProduct, Long idUser) {
@@ -64,7 +74,6 @@ public class ReviewRepositoryImplementation implements ReviewRepository {
             User user = session.get(User.class, idUser);
             Hibernate.initialize(user.getReviews());
 
-            System.out.println(product);
             product.addReview(review);
             user.addReview(review);
